+++ packages/graphql-generator/src/__tests__/models.test.ts
@@ -1,16 +1,25 @@
-import { generateModels, GenerateModelsOptions, ModelsTarget } from '..';
+import { generateModels, generateModelsSync, GenerateModelsOptions, ModelsTarget } from '..';
 import { readSchema } from './utils';
 
-describe('generateModels', () => {
-  describe('targets', () => {
-    const targets: ModelsTarget[] = ['java', 'swift', 'javascript', 'typescript', 'dart', 'introspection'];
-    targets.forEach(target => {
-      test(`basic ${target}`, async () => {
+const generators = ['generateModels', 'generateModelsSync'] as const;
+const targets: ModelsTarget[] = ['java', 'swift', 'javascript', 'typescript', 'dart', 'introspection'];
+
+async function runGenerator(generator: (typeof generators)[number], options: GenerateModelsOptions) {
+  if (generator === 'generateModels') {
+    return await generateModels(options);
+  } else {
+    return generateModelsSync(options);
+  }
+}
+
+describe.each(generators)('generateModels with', (generator) => {
+    describe('targets', () => {
+      test.each(targets)(`basic`, async (target) => {
         const options: GenerateModelsOptions = {
           schema: readSchema('blog-model.graphql'),
           target,
         };
-        const models = await generateModels(options);
+        const models = await runGenerator(generator, options);
         expect(models).toMatchSnapshot();
       });
     });
@@ -21,25 +30,36 @@ describe('generateModels', () => {
         target: 'swift',
         improvePluralization: true,
       };
-      const models = await generateModels(options);
+      const models = await runGenerator(generator, options);
+      expect(models).toMatchSnapshot();
+    });
+
+    test('does not fail on custom directives', async () => {
+      const options: GenerateModelsOptions = {
+        schema: `
+          type Blog @customModel {
+            id: ID!
+            name: String! @customField
+          }`,
+        target: 'introspection',
+        directives: `
+          directive @customModel on OBJECT
+          directive @customField on FIELD_DEFINITION
+        `,
+      };
+      const models = await runGenerator(generator, options);
       expect(models).toMatchSnapshot();
     });
   });
 
-  test('does not fail on custom directives', async () => {
+targets.forEach(target => {
+  test(`both generates generate the same basic ${target}`, async () => {
     const options: GenerateModelsOptions = {
-      schema: `
-        type Blog @customModel {
-          id: ID!
-          name: String! @customField
-        }`,
-      target: 'introspection',
-      directives: `
-        directive @customModel on OBJECT
-        directive @customField on FIELD_DEFINITION
-      `,
+      schema: readSchema('blog-model.graphql'),
+      target,
     };
-    const models = await generateModels(options);
-    expect(models).toMatchSnapshot();
+    const asyncGeneratorModels = await generateModels(options);
+    const syncGeneratorModels = generateModelsSync(options);
+    expect(asyncGeneratorModels).toEqual(syncGeneratorModels);
   });
 });